requirements.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
sqlalchemy
